tentativa do ajuste do erro

+++ Exemplo Spring Aula - Alunos/Exemplo Spring Aula - Alunos/demo/src/main/java/com/example/demo/Entities/Destino.java
@@ -33,6 +33,6 @@ public class Destino {
     @Column(nullable = false)
     private String localizacao;
     @Column(nullable = false)
-    private float precoPorPessoa;
+    private Double precoPorPessoa;
 
 }
